add via for cnc reference and replace mount pads
Furthermore bump version to v2.3 because we already etched v2.2 and
changed the board afterwards so e.g. milling some v2.2 is not possible
with current master anymore. There's still no tag for v2.2 but I guess
there just won't be any, maybe a branch must be enough for now.